[ticket/12466] Move classes from acp_database.php to their own files
 * Moving classes from acp_database.php to phpbb/db/extractor namespace,
   also into separate files
 * Adding DocBlocks and property visibility to classes
 * Removing globals from code
 * Passing former globals to base_extractor's constructor
 * Adding DB extractor as a service, also implementing the extractor interface
   as well as the extractor factory.

PHPBB3-12466
